feat: update LabelsData interface to OSD6 data pipeline
