+++ packages/@tissuumaps-core/src/types/openseadragon.ts
@@ -1,5 +1,21 @@
 import type OpenSeadragon from "openseadragon";
 
+/** Configuration object accepted by OpenSeadragon as a tile source */
+export type TileSourceConfig = object;
+
+/** A custom tile source that resolves tile URLs programmatically */
+export interface CustomTileSource {
+  /**
+   * Returns the URL for a specific tile
+   *
+   * @param level - Pyramid level
+   * @param x - Tile column index
+   * @param y - Tile row index
+   * @returns The tile URL, or a function that returns the URL
+   */
+  getTileUrl(level: number, x: number, y: number): string | (() => string);
+}
+
 /**
  * OpenSeadragon options, as consumed by the `OpenSeadragonContext` of
  * `@tissuumaps/render`
@@ -67,44 +83,4 @@ export type OpenSeadragonViewerOptions = Omit<
 
   /** Unsupported */
   referenceStripElement?: never;
-};
-
-/**
- * Options for the WebGL renderers
- */
-export type WebGLOptions = {
-  /** Options for rendering the project's points */
-  pointsRenderOptions: WebGLPointsRenderOptions;
-
-  /** Options for rendering the project's shapes */
-  shapesRenderOptions: WebGLShapesRenderOptions;
-};
-
-/**
- * Options for rendering points
- */
-export type WebGLPointsRenderOptions = {
-  /**
-   * Unitless factor by which the size of every point is multiplied
-   *
-   * Applies on top of the layer- and object-level point size settings.
-   */
-  globalPointSizeFactor: number;
-};
-
-/**
- * Options for rendering shapes
- */
-export type WebGLShapesRenderOptions = {
-  /** Width of the shape outlines, in world coordinates */
-  strokeWidth: number;
-
-  /**
-   * Number of horizontal scanlines a shapes object's bounding box is divided into
-   *
-   * For each scanline, the rasterizer precomputes which polygon edges cross it,
-   * so that the fragment shader only has to consider those. More scanlines mean
-   * fewer edges to test per fragment, at the cost of a larger scanline texture.
-   */
-  numScanlines: number;
 };
